Add separate scroll gesture handling for MacOS
MacOS issues the same event for a mouse scroll wheel as for a two-finger
scroll gesture on a trackpad.  However, these can be differentiated
by testing the event subtype == NSEventSubtypeMouseEvent.

In addition, the magnitudes of deltaX and deltaY are too small when
returned by a trackpad scroll, and scrollingDeltaX and
scrollingDeltaY should be used instead.

This introduces a new FL_SCROLL_GESTURE to enable FLTK applications to
use this gesture.