Improve macOS touch gestures -- experimental + debug
Output gesture events, add total zoom and total rotation values.
The total values are intended to help testing. Each total value
is reset to 1.0 and 0.0, respectively, when the "other" gesture
is processed.

Fix zoom value by adding 1.0 as documented in macOS docs to make the
value usable as a zoom factor.

Multiply zoom and rotation with 100,000 rather than 1,000 for better
granularity.

This is still a proof of concept used for development purposes.