Implement somewhat more sophisticated widget based on Fl_Table
As compared to Fl_Table, this widget does own the data -- as a pointer
to the vector <T> or vector<vector<T> >.

It can copy/paste blocks of data both internally and to tools like Excel.

If pasted data extends beyond the size of the table, the table will
be expanded to fit.  There is are options to disable this behavior
for rows and columns -- if expansion is attempted but disabled, the system
will beep and do nothing.